+++ src/App.js
@@ -1,8 +1,8 @@
+import { Routes, Route } from 'react-router-dom';
 import StartPage from './pages/StartPage';
 import LoadingIntroPage from './pages/LoadingIntroPage';
 import LoadingTransitionPage from './pages/LoadingTransitionPage';
 import ResultPage from './pages/ResultPage';
-import { Routes, Route } from 'react-router-dom';
 
 function App() {
   return (
