+++ src/font.css
@@ -1,4 +1,4 @@
 @font-face {
   font-family: 'Pretendard';
-  src: url('assets/fonts/PretendardVariable.ttf') format('truetype');
+  src: url('fonts/PretendardVariable.ttf') format('truetype');
 }
