+++ src/pages/LoadingTransitionPage.jsx
@@ -5,7 +5,7 @@ import { useNavigate } from 'react-router-dom';
 import CommonLayoutSize from '../components/CommonLayout.jsx';
 import { BounceLoader } from 'react-spinners';
 
-import { TRANSITION_MESSAGE } from '../constants/loading.js';
+import { TRANSITION_MESSAGE } from '../constants/loadingSettings.js';
 
 function LoadingTransitionPage() {
   const navigate = useNavigate();
